Tools for convergence
method needs to be cleaned, lots of unused functions. The only ones of interest are cma_energy and opt_res_c7. The first one uses CMA and the second one is a double objective NEWUOA (you can change wheter you want it to be a dual objective or not by changing a coefficient when calling the function, usually I call this coefficient "double".