3D Script
Works BUT, there are some complications : 
Mainly concerning the regularization, I'm using a Thikonov regularization when assembling the full 3D matrix from the elemental one (K_local). The problem is that without it, the matrix can be singular or ill-conditioned so this system can be hard to solve, especially with iterative solvers. This regularization aims to solve this problem. But this solution has a cost, the regularization is linked to a coefficient which is itself linked to the system we're analyzing. Meaning we have to change it if we change the conditions/size of our problem (it's not that difficult to change but we have to find some kind of predictive way to change it or to find the sweet spot for all sizes and conditions).
If the matrix is ill-conditioned, the optimization behaves weirdly and most of the vector state is composed in the end of nan values, causing the energy to be even lower than the reference point.